This is the primary game loop
For now all of the individual functions, like the pet menu and shop, are in this program as well, but they can be separated so we don't have to all be working on the same file. If you decide to work on one of them then probably try and separate it into a different cpp file so no one accidentally overrides someone else's work.